+++ src/js/api.js
@@ -1,4 +1,4 @@
-class Api {
+export class Api {
     constructor(options) {
         this.baseUrl = options.baseUrl; 
         this.headers = options.headers;
